Works now to an extent...
Updated the UI and added the Subsequence class. It now can load an fna
file, let the user enter a DNA sequence and try to match it and its
edit1 distances to the file. The program becomes unresponsive while
trawling the file. A file of 40MB took 5-10 minutes for me.